delete db시간 기준으로 변경 테스트 2개 고장남

+++ src/main/java/com/park/monitoring/service/ServerInfoService.java
@@ -5,6 +5,7 @@ import com.park.monitoring.model.ServerInfo;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -55,6 +56,9 @@ public List<ServerInfo> findAllServerInfo() {
             || serverInfo.getMemoryTotal() == null) {
             throw new IllegalArgumentException("필수 값이 null입니다.");
         }
+        if(serverInfoMapper.isIpExists(serverInfo.getServerIp()) == 1){
+            throw new DataIntegrityViolationException("이미 존재하는 ip입니다.");
+        }
         int result = serverInfoMapper.insertServerInfo(serverInfo);
         if (result < 1) throw new RuntimeException("데이터 등록 실패");
         return result;
@@ -68,8 +72,8 @@ public List<ServerInfo> findAllServerInfo() {
                 || serverInfo.getServerIp() == null) {
             throw new IllegalArgumentException("입력받은 값이 비정상입니다. - " + this.getClass());
         }
-        if (serverInfo == null) {
-            throw new NoSuchElementException("수정할 Server 데이터를 찾을수 없습니다. - " + this.getClass());
+        if (serverInfoMapper.selectServerInfoById(serverInfo.getServerId()) == null) {
+            throw new NoSuchElementException("수정할 Server를 찾을수 없습니다. - " + this.getClass());
         }
         int result = serverInfoMapper.updateServerInfo(serverInfo);
         if (result < 1) {
@@ -83,6 +87,9 @@ public List<ServerInfo> findAllServerInfo() {
         if (id == null) {
             throw new IllegalArgumentException("입력된 Id가 null입니다.");
         }
+        if(serverInfoMapper.selectServerInfoById(id)==null){
+            throw new NoSuchElementException("존재하지 않는 서버입니다.");
+        }
         int result = serverInfoMapper.deleteServerInfoById(id);
         if (result < 1) throw new RuntimeException("데이터 삭제 실패");
         else return result;
@@ -93,7 +100,7 @@ public List<ServerInfo> findAllServerInfo() {
         int result = serverInfoMapper.deleteAll();
         if (result >= 1) return result;
         else {
-            throw new DataIntegrityViolationException("데이터 삭제 실패");
+            throw new EmptyResultDataAccessException("삭제할 데이터가 존재하지 않습니다.",1);
         }
     }
 }
